refactor(tasks): extract mode handlers to prompts/ modules
